requirements.txt
opencv-python
pyautogui
numpy
mss
screeninfo
black
isort
flake8
matplotlib
pillow
pytesseract
pyqt5